use trait to define checkoutSparse and checkout only once
