+++ src/main/scala/io/viash/functionality/dependencies/AbstractGitRepository.scala
@@ -24,6 +24,15 @@ import java.nio.file.Paths
 
 abstract class AbstractGitRepository extends Repository {
   val uri: String
+  val fullUri: String
+
+  def copyRepo(
+    name: String,
+   `type`: String,
+    tag: Option[String],
+    path: Option[String],
+    localPath: String
+  ): AbstractGitRepository
 
   def hasBranch(name: String, cwd: Option[File]): Boolean = {
     val out = Exec.runCatch(
@@ -40,7 +49,55 @@ abstract class AbstractGitRepository extends Repository {
     )
     out.exitValue == 0
   }
-
+  
+  
   // Get the repository part of where dependencies should be located in the target/dependencies folder
   def subOutputPath: String = Paths.get(`type`, uri, tag.getOrElse("")).toString()
+
+  // Clone of single branch with depth 1 but without checking out files
+  def checkoutSparse(): AbstractGitRepository = {
+    val temporaryFolder = IO.makeTemp("viash_hub_repo")
+    val cwd = Some(temporaryFolder.toFile)
+
+    Console.err.println(s"temporaryFolder: $temporaryFolder uri: $uri fullUri: $fullUri")
+
+    val singleBranch = tag match {
+      case None => List("--single-branch")
+      case Some(value) => List("--single-branch", "--branch", value)
+    }
+
+    val loggers = Seq[String => Unit] { (str: String) => {Console.err.println(str)} }
+
+    val out = Exec.runCatch(
+      List("git", "clone", fullUri, "--no-checkout", "--depth", "1") ++ singleBranch :+ ".",
+      cwd = cwd,
+      loggers = loggers,
+    )
+
+    copyRepo(localPath = temporaryFolder.toString)
+  }
+
+  // Checkout of files from already cloned repository. Limit file checkout to the path that was specified
+  def checkout(): AbstractGitRepository = {
+    val pathStr = path.getOrElse(".")
+    val cwd = Some(Paths.get(localPath).toFile)
+    val checkoutName = tag match {
+      case Some(name) if hasBranch(name, cwd) => s"origin/$name"
+      case Some(name) if hasTag(name, cwd) => s"tags/$name"
+      case _ => "origin/HEAD"
+    }
+
+    val out = Exec.runCatch(
+      List("git", "checkout", checkoutName, "--", pathStr),
+      cwd = cwd
+    )
+
+    Console.err.println(s"checkout out: ${out.command} ${out.exitValue} ${out.output}")
+
+    if (path.isDefined)
+      copyRepo(localPath = Paths.get(localPath, path.get).toString)
+    else
+      // no changes to be made
+      this
+  }
 }
