+++ src/main/scala/io/viash/functionality/dependencies/Repository.scala
@@ -24,7 +24,7 @@ import java.nio.file.Paths
 import java.nio.file.Files
 
 @description("Specifies a repository where dependency components can be found.")
-abstract class Repository {
+abstract class Repository extends CopyableRepo[Repository] {
   @description("The identifier used to refer to this repository from dependencies.")
   val name: String
